+++ CHANGELOG.md
@@ -14,10 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added kql.Builder struct for safe building of KQL statements from variables without use of 'Unsafe' mode.
 * Simpler handling of query parameters using kql.Parameters struct.
+* TokenCredential support for authentication.
 ### Security
 * No redirects are allowed by default.
 ### Fixed
 * Replace non-ascii characters in headers to be in line with the service.
+* DefaultCredential now uses the same HTTP client as the rest of the SDK.
 
 ## [0.11.3] - 2023-03-20
 ### Added